agents/skill_mastery_agent.py
"""
Skill Mastery Agent - Google ADK version.
Analyzes student proficiency across subjects using Gemini LLM.
"""
from typing import Dict, Any
import os

try:
    from google.adk import Agent
    ADK_AVAILABLE = True
except ImportError:
    ADK_AVAILABLE = False
    print("Warning: google-adk not installed. Using fallback mode.")


def run_skill_mastery(student_id: str) -> Dict[str, Any]:
    """
    Analyze student skill levels across subjects.
    Fallback function when ADK not available.
    
    Args:
        student_id: The student's ID
        
    Returns:
        Dictionary with skill profile
    """
    # Mock data for fallback
    skill_profile = {
        "Math": "needs_support",
        "ELA": "on_track",
        "Science": "strong"
    }
    return {"skill_profile": skill_profile}



